@@ -17,8 +17,9 @@
   "packageManager": "yarn@3.6.0",
   "scripts": {
     "serve": "run-p 'watch:**'",
-    "build": "REBUILD_VERSION_CACHE=true run-s build-dev:scripts 'build:**'",
-    "build-prod": "NETLIFY=true run-s build-prod:scripts 'build:**'",
+    "build": "run-s clean-version-cache build-dev:scripts 'build:**'",
+    "clean-version-cache": "rm -f source/_data/versionCache.json",
+    "build-prod": "NETLIFY=true run-s clean-version-cache build-prod:scripts 'build:**'",
     "build:sass": "sass --style=compressed ./source/assets/sass/sass.scss:./source/assets/dist/css/sass.css ./source/assets/sass/noscript.scss:./source/assets/dist/css/noscript.css",
     "watch:sass": "sass --watch ./source/assets/sass/sass.scss:./source/assets/dist/css/sass.css ./source/assets/sass/noscript.scss:./source/assets/dist/css/noscript.css",
     "build-dev:scripts": "rollup -c",
@@ -36,7 +37,9 @@
     "stylelint:ci": "stylelint '**/*.{css,scss}'",
     "lint": "run-s prettier gts stylelint tsc",
     "lint:ci": "run-s prettier:ci gts:ci stylelint:ci tsc",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "checklinks:internal": "npx --yes --package=hyperlink --package=tap-spot -- 'hyperlink --canonicalroot https://sass-lang.com/ --root _site/ -r -p -i --skip \".css.map\" --skip \".js.map\" --skip \".svg\" --skip \"feed.atom\" --skip \"/documentation/js-api/\" --skip \"/blog/\" _site/index.html | tap-spot'",
+    "checklinks:external": "npx --yes --package=broken-link-checker -- 'blc -ro --exclude=localhost:8080 --exclude=codepen.io --exclude=linkedin.com --exclude=docs.github.com --exclude=twitter.com https://sass-lang.com'"
   },
   "devDependencies": {
     "@11ty/eleventy": "^2.0.1",
@@ -47,11 +50,12 @@
     "@babel/preset-typescript": "^7.22.5",
     "@codemirror/lang-sass": "^6.0.1",
     "@rollup/plugin-babel": "^6.0.3",
-    "@rollup/plugin-commonjs": "^25.0.1",
+    "@rollup/plugin-commonjs": "^25.0.2",
     "@rollup/plugin-inject": "^5.0.3",
     "@rollup/plugin-node-resolve": "^15.1.0",
     "@rollup/plugin-terser": "^0.4.3",
     "@sindresorhus/slugify": "^1.1.2",
+    "@types/deep-equal": "^1.0.1",
     "@types/jquery": "^3.5.16",
     "@types/jqueryui": "^1.12.17",
     "@types/lodash": "^4.14.195",
@@ -60,13 +64,13 @@
     "@types/markdown-it-footnote": "^3.0.0",
     "@types/node": "^16",
     "@types/prismjs": "^1.26.0",
-    "@typescript-eslint/eslint-plugin": "^5.59.9",
-    "@typescript-eslint/parser": "^5.59.9",
+    "@typescript-eslint/eslint-plugin": "^5.60.0",
+    "@typescript-eslint/parser": "^5.60.0",
     "cheerio": "^1.0.0-rc.12",
     "codemirror": "^6.0.1",
     "date-fns": "^2.30.0",
     "deep-equal": "^2.2.1",
-    "eslint": "^8.42.0",
+    "eslint": "^8.43.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-import-resolver-typescript": "^3.5.5",
     "eslint-plugin-import": "^2.27.5",
@@ -89,10 +93,10 @@
     "prettier": "^2.8.8",
     "prismjs": "^1.29.0",
     "rollup": "^3.25.1",
-    "sass": "^1.63.3",
-    "semver-parser": "^4.1.4",
+    "sass": "^1.63.4",
+    "semver": "^7.5.2",
     "strip-indent": "^3.0.0",
-    "stylelint": "^15.7.0",
+    "stylelint": "^15.8.0",
     "stylelint-config-standard-scss": "^9.0.0",
     "truncate-html": "^1.0.4",
     "ts-node": "^10.9.1",
